Added start/stop mojos.

+++ vbox-core/src/main/java/com/alexecollins/util/FileUtils2.java
@@ -18,12 +18,15 @@ public class FileUtils2 {
 	 * @see org.apache.commons.io.FileUtils#copyURLToFile(java.net.URL, java.io.File)
 	 */
 	public static void copyURLToFile(URL url, File file) throws IOException {
+		final File ok = new File(file + ".ok");
+		if (ok.exists()) {return;}
 		URLConnection connection = url.openConnection();
 		if (connection instanceof HttpURLConnection) {
 			getHttpUrl(file, (HttpURLConnection)connection);
 		} else {
 			FileUtils.copyURLToFile(url, file);
 		}
+		FileUtils.writeStringToFile(ok, "marker file that indicates " + file + " downloaded OK");
 	}
 
 	private static void getHttpUrl(File file, HttpURLConnection connection) throws IOException {
@@ -32,15 +35,7 @@ public class FileUtils2 {
 		connection.setRequestProperty("Range", "bytes=" + downloaded + "-");
 		connection.setDoInput(true);
 		connection.setDoOutput(true);
-		final BufferedInputStream in;
-		try {
-			in = new BufferedInputStream(connection.getInputStream());
-		} catch (IOException e) {
-			// range invalid, we've got the whole file (or an almighty f**kup)
-			if (!e.getMessage().contains("HTTP response code: 416"))
-				throw e;
-			else return;
-		}
+		final BufferedInputStream in = new BufferedInputStream(connection.getInputStream());
 		try {
 			final BufferedOutputStream out = new BufferedOutputStream(new FileOutputStream(file, true), 1024);
 			try {
